requirements.txt
## python=3.10.14
accelerate=0.21.0
aiofiles=23.2.1
aiohttp=3.9.3
aiosignal=1.3.1
altair=5.2.0
antlr4-python3-runtime=4.9.3
anyio=4.3.0
appdirs=1.4.4
async-timeout=4.0.3
attrs=23.2.0
av=12.0.0
bitsandbytes=0.41.0
braceexpand=0.1.7
brotli=1.1.0
cffi=1.16.0
charset-normalizer=3.3.2
click=8.1.7
clip=1.0
cmake=3.29.0.1
contourpy=1.2.0
cycler=0.12.1
datasets=2.18.0
decorator=4.4.2
decord=0.6.0
diffusers=0.27.2
dill=0.3.8
docker-pycreds=0.4.0
docopt=0.6.2
einops=0.6.1
einops-exts=0.0.4
exceptiongroup=1.2.0
fastapi=0.110.0
ffmpeg-python=0.2.0
ffmpy=0.3.2
fidelity=0.1
filelock=3.13.3
fire=0.4.0
fonttools=4.50.0
frozenlist=1.4.1
fsspec=2024.2.0
ftfy=6.2.0
future=1.0.0
fvcore=0.1.5.post20221221
gitdb=4.0.11
gitpython=3.1.42
gradio=3.35.2
gradio-client=0.2.9
h11=0.14.0
httpcore=0.17.3
httpx=0.24.0
huggingface-hub=0.22.1
idna=3.6
imageio=2.34.0
imageio-ffmpeg=0.4.9
importlib-metadata=7.1.0
iopath=0.1.10
jinja2=3.1.3
joblib=1.3.2
jsonschema=4.21.1
lit=18.1.2
llava=1.1.3
lpips=0.1.4
madgrad=1.3
markdown-it-py=2.2.0
markdown2=2.4.13
markupsafe=2.1.5
matplotlib=3.8.3
mdit-py-plugins=0.3.3
mdurl=0.1.2
moviepy=1.0.3
mpmath=1.3.0
multidict=6.0.5
multiprocess=0.70.16
mutagen=1.47.0
ncurses=6.4=h6a678d5_0
networkx=3.2.1
numpy=1.26.4
nvidia-cublas-cu11=11.10.3.66
nvidia-cuda-cupti-cu11=11.7.101
nvidia-cuda-nvrtc-cu11=11.7.99
nvidia-cuda-runtime-cu11=11.7.99
nvidia-cudnn-cu11=8.5.0.96
nvidia-cufft-cu11=10.9.0.58
nvidia-curand-cu11=10.2.10.91
nvidia-cusolver-cu11=11.4.0.1
nvidia-cusparse-cu11=11.7.4.91
nvidia-nccl-cu11=2.14.3
nvidia-nvtx-cu11=11.7.91
omegaconf=2.3.0
opencv-python=4.9.0.80
openssl=3.0.13
orjson=3.10.0
packaging=24.0
pandas=2.2.1
parameterized=0.9.0
peft=0.4.0
pillow=10.2.0
portalocker=2.8.2
proglog=0.1.10
protobuf=4.25.3
psutil=5.9.8
pyarrow=15.0.2
pyarrow-hotfix=0.6
pycparser=2.21
pycryptodomex=3.20.0
pydantic=1.10.14
pydub=0.25.1
pygments=2.17.2
pyparsing=3.1.2
pytorch-fid=0.3.0
pytorchvideo=0.1.5
pytube=15.0.0
pytz=2024.1
pyyaml=6.0.1
readline=8.2
referencing=0.34.0
regex=2023.12.25
requests=2.31.0
rpds-py=0.18.0
safetensors=0.4.2
scenedetect=0.6
scikit-learn=1.2.2
scipy=1.12.0
segment-anything=1.0
semantic-version=2.10.0
sentencepiece=0.1.99
sentry-sdk=1.44.0
setproctitle=1.3.3
setuptools=68.2.2
shortuuid=1.0.13
six=1.16.0
smmap=5.0.1
sniffio=1.3.1
soundfile=0.12.1
sqlite=3.41.2
starlette=0.36.3
svgwrite=1.4.3
sympy=1.12
tabulate=0.9.0
termcolor=2.4.0
threadpoolctl=3.4.0
timeout-decorator=0.5.0
timm=0.6.13
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3
toolz=0.12.1
torch=2.0.1
torch-fidelity=0.3.0
torchaudio=2.0.2
torchdata=0.7.1
torchvision=0.15.2
tqdm=4.66.2
transformers=4.31.0
triton=2.0.0
typing-extensions=4.10.0
tzdata=2024.1
uc-micro-py=1.0.3
urllib3=2.2.1
uvicorn=0.29.0
wandb=0.16.5
wavedrom=2.0.3.post3
wcwidth=0.2.13
webdataset=0.2.86
websockets=12.0
webvtt-py=0.4.6
wheel=0.41.2
xxhash=3.4.1
xz=5.4.6
yacs=0.1.8
yarl=1.9.4
zipp=3.18.1
zlib=1.2.13
